Moved compoinent files to related valueType directories
