+++ src/valueTypes/BooleanValueType/component.js
