+++ src/valueTypes/FloatValueType/component.js
